requirements.txt
Django==1.11.7
django-autoslug==1.9.3
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
PyJWT==1.5.3
python-decouple==3.1
pytz==2017.3
